# tests: improve tests with a unified `env.Reset`