--- a/tests/server/api/utils.go
+++ b/tests/server/api/utils.go
@@ -15,15 +15,10 @@
 package api
 
 import (
-	"encoding/json"
 	"fmt"
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/pingcap/kvproto/pkg/metapb"
-
-	"github.com/tikv/pd/pkg/errs"
-	"github.com/tikv/pd/pkg/schedule/placement"
 	"github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/tests"
 )
@@ -39,62 +34,5 @@ func pauseAllCheckers(re *require.Assertions, cluster *tests.TestCluster) {
 		err = testutil.ReadGetJSON(re, tests.TestDialClient, url, &resp)
 		re.NoError(err)
 		re.True(resp["paused"].(bool))
-	}
-}
-
-func cleanStoresAndRegions(re *require.Assertions) func(*tests.TestCluster) {
-	return func(cluster *tests.TestCluster) {
-		// clean region cache
-		for _, server := range cluster.GetServers() {
-			pdAddr := cluster.GetConfig().GetClientURL()
-			for _, region := range server.GetRegions() {
-				url := fmt.Sprintf("%s/pd/api/v1/admin/cache/region/%d", pdAddr, region.GetID())
-				err := testutil.CheckDelete(tests.TestDialClient, url, testutil.StatusOK(re))
-				re.NoError(err)
-			}
-			re.Empty(server.GetRegions())
-		}
-
-		// clean stores
-		leader := cluster.GetLeaderServer()
-		for _, store := range leader.GetStores() {
-			if store.NodeState == metapb.NodeState_Removed {
-				continue
-			}
-			err := cluster.GetLeaderServer().GetRaftCluster().RemoveStore(store.GetId(), true)
-			if err != nil {
-				re.ErrorIs(err, errs.ErrStoreRemoved)
-			}
-			re.NoError(cluster.GetLeaderServer().GetRaftCluster().BuryStore(store.GetId(), true))
-		}
-		re.NoError(cluster.GetLeaderServer().GetRaftCluster().RemoveTombStoneRecords())
-		re.Empty(leader.GetStores())
-		testutil.Eventually(re, func() bool {
-			if sche := cluster.GetSchedulingPrimaryServer(); sche != nil {
-				for _, s := range sche.GetBasicCluster().GetStores() {
-					if s.GetState() != metapb.StoreState_Tombstone {
-						return false
-					}
-				}
-			}
-			return true
-		})
-	}
-}
-
-func cleanRules(re *require.Assertions) func(*tests.TestCluster) {
-	return func(cluster *tests.TestCluster) {
-		// clean rules
-		defaultRule := placement.GroupBundle{
-			ID: "pd",
-			Rules: []*placement.Rule{
-				{GroupID: "pd", ID: "default", Role: "voter", Count: 3},
-			},
-		}
-		data, err := json.Marshal([]placement.GroupBundle{defaultRule})
-		re.NoError(err)
-		urlPrefix := cluster.GetLeaderServer().GetAddr()
-		err = testutil.CheckPostJSON(tests.TestDialClient, urlPrefix+"/pd/api/v1/config/placement-rule", data, testutil.StatusOK(re))
-		re.NoError(err)
 	}
 }