<!--

Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/tikv/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

<!--

Please create an issue first to describe the problem.
There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: Ref #4399

### What is changed and how does it work?

avoid unnecessarily creating and destroying clusters and use a unified `env.Reset`

Status | Chunk | Time | Link
-- | -- | -- | --
Before Optimization | chunk5 | 7min 18s | [Job 46535627785](https://github.com/tikv/pd/actions/runs/16463556492/job/46535627785)
Optimized | chunk5 | 6min 1s | [Job 46564645823](https://github.com/tikv/pd/actions/runs/16472366678/job/46564645823)
Before Optimization | chunk6 | 6min 37s | [Job 46535627798](https://github.com/tikv/pd/actions/runs/16463556492/job/46535627785)
Optimized | chunk6 | 5min 30s | [Job 46564645831](https://github.com/tikv/pd/actions/runs/16472366678/job/46564645831)



<!--

You could use the "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#format-of-the-commit-message.

-->

```commit-message
```

### Check List

<!-- Remove the items that are not applicable. -->

Tests

<!-- At least one of these tests must be included. -->

- Unit test


### Release note

<!--

A bugfix or a new feature needs a release note. If there is no need to give a release note, just leave it with the `None`.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

-->

```release-note
None.
```